feat: 🔥 add any to string & refactor
